fixed loyalty status

+++ src/Decorator/LoyaltyStatus.java
@@ -2,9 +2,9 @@ package Decorator;
 
 public class LoyaltyStatus {
 
-    double customer_discount = 0;
+    int customer_discount = 0;
 
-    public double discount(String discount){
+    public int discount(String discount){
         if(discount == "basic"){
             customer_discount = 1;
             System.out.println("Applying discout of $" + customer_discount);
